+++ apps/terasu-web/src/app/(authenticated)/companies/[Id]/components/details/yearly-info.tsx
@@ -10,7 +10,7 @@ import { BasicInfoSection } from "./sections/basic-info-section/basic-info-secti
 import { BranchSection } from "./sections/branch-section";
 import { ContactSection } from "./sections/contact-section";
 import { PhilosophySection } from "./sections/philosophy-section/philosophy-info";
-import { WelfareSection } from "./sections/welfare-section";
+import { WelfareSection } from "./sections/welfare-info";
 import type { CacheData, YearDateData, YearlyDetailItem } from "./types";
 import { YearlyInfoTabs } from "./yearly-info-tabs";
